configure swagger and user auth

+++ src/speakers/speakers.controller.ts
@@ -11,6 +11,7 @@ import {
   Render,
   Req,
   UploadedFile,
+  UseGuards,
   UseInterceptors,
 } from '@nestjs/common';
 import { SpeakersService } from './speakers.service';
@@ -18,31 +19,66 @@ import { Request } from 'express';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { MinioService } from 'src/minio/minio.service';
 import { CreateSpeakerDto } from './dto/create-speaker.dto';
+import {
+  ApiBody,
+  ApiConsumes,
+  ApiOperation,
+  ApiQuery,
+  ApiResponse,
+  ApiTags,
+} from '@nestjs/swagger';
+import {
+  BadRequestStatusType,
+  NotFoundStatusType,
+  SpeakerCreateBody,
+  SpeakerUpdateBody,
+  SpeakersResponseType,
+} from 'src/types';
+import { Speaker } from './speakers.model';
+import { Meetup } from 'src/meetups/meetups.model';
+import { JwtService } from '@nestjs/jwt';
+import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
+import { RolesGuard } from 'src/auth/roles.guard';
+import { Roles } from 'src/auth/roles-auth.decorator';
+import defaultAvatar from '../assets/defaultAvatar.jpg'
 
+@ApiTags('Спикеры')
 @Controller('speakers')
 export class SpeakersController {
   constructor(
     private speakersService: SpeakersService,
     private readonly minioService: MinioService,
+    private jwtService: JwtService,
   ) {}
 
+  @ApiOperation({ summary: 'Получить спикеров по компании' })
+  @ApiResponse({ status: 200, type: SpeakersResponseType })
+  @ApiQuery({
+    type: 'string',
+    name: 'company',
+    required: false,
+  })
   @Get('')
-  @Render('SpeakersPage')
-  async getByStatus(@Req() request?: Request) {
-    const userID = 1;
+  //@Render('SpeakersPage')
+  async getByCompany(@Req() request?: Request) {
+    const token = request.cookies.meetups_access_token?.token;
+
+    const user = token && this.jwtService.verify(token, { secret: 'SECRET' });
 
     let company = request.query.company?.toString();
 
     let speakers = await this.speakersService.getByOrganization(
-      userID,
+      user?.id,
       company,
     );
 
     return { ...speakers, company };
   }
 
+  @ApiOperation({ summary: 'Получить спикера по id' })
+  @ApiResponse({ status: 200, type: Speaker })
   @Get(':id')
-  @Render('SingleSpeakerPage')
+  //@Render('SingleSpeakerPage')
   getById(@Param('id', ParseIntPipe) id: number) {
     return this.speakersService.getById(id);
   }
@@ -53,6 +89,12 @@ export class SpeakersController {
   //    return this.speakersService.changeStatus(id);
   //  }
 
+  @ApiOperation({ summary: 'Создать нового спикера' })
+  @ApiConsumes('multipart/form-data')
+  @ApiBody({ type: SpeakerCreateBody })
+  @ApiResponse({ status: 201, type: Speaker })
+  @Roles('модератор')
+  @UseGuards(RolesGuard)
   @Post()
   @UseInterceptors(FileInterceptor('file'))
   async create(
@@ -72,6 +114,22 @@ export class SpeakersController {
     return speaker;
   }
 
+  @ApiOperation({ summary: 'Изменить фотографию спикера' })
+  @ApiConsumes('multipart/form-data')
+  @ApiBody({
+    schema: {
+      type: 'object',
+      properties: {
+        file: {
+          type: 'string',
+          format: 'binary',
+        },
+      },
+    },
+  })
+  @ApiResponse({ status: 200, type: Speaker })
+  @Roles('модератор')
+  @UseGuards(RolesGuard)
   @Put('/image/:id')
   @UseInterceptors(FileInterceptor('file'))
   async uploadImage(
@@ -84,41 +142,80 @@ export class SpeakersController {
       currentSpeaker.avatarImg.split('meetups-app')[1],
     );
 
-    await this.speakersService.uploadAvatar(id, file.originalname);
-
-    const fileName = await this.minioService.uploadFile(
-      `/speakers/${id}`,
-      file,
+    const updatedSpeaker = await this.speakersService.uploadAvatar(
+      id,
+      file.originalname,
     );
 
-    return `Изображение изменено на ${fileName}`;
+    this.minioService.uploadFile(`/speakers/${id}`, file);
+
+    return updatedSpeaker;
   }
 
+  @ApiOperation({ summary: 'Добавить спикера в митап' })
+  @ApiResponse({ status: 200, type: Meetup })
+  @ApiResponse({ status: 400, type: BadRequestStatusType })
+  @UseGuards(JwtAuthGuard)
   @Post(':id')
-  addToMeetup(@Param('id', ParseIntPipe) id: number) {
-    let userID = 1;
-    return this.speakersService.addSpeakerToMeetup(id, userID);
+  addToMeetup(@Param('id', ParseIntPipe) id: number, @Req() request: Request) {
+    const token = request.cookies.meetups_access_token.token;
+    const user = this.jwtService.verify(token, { secret: 'SECRET' });
+    return this.speakersService.addSpeakerToMeetup(id, user.id);
   }
 
+  @ApiOperation({ summary: 'Удалить спикера' })
+  @ApiResponse({ status: 200, type: SpeakersResponseType })
+  @ApiResponse({ status: 404, type: NotFoundStatusType })
+  @Roles('модератор')
+  @UseGuards(RolesGuard)
   @Delete(':id')
-  async delete(@Param('id', ParseIntPipe) id: number) {
-    const userID = 1;
+  async delete(@Param('id', ParseIntPipe) id: number, @Req() request: Request) {
+    const token = request.cookies.meetups_access_token.token;
+    const user = this.jwtService.verify(token, { secret: 'SECRET' });
 
     const speaker = await this.getById(id);
 
-    if (speaker)
+    if (speaker.avatarImg && !speaker.avatarImg.includes('/default/defaultAvatar.jpg'))
       await this.minioService.deleteFile(
         speaker.avatarImg.split('meetups-app')[1],
       );
 
-    return this.speakersService.deleteSpeaker(userID, id);
+    return this.speakersService.deleteSpeaker(user.id, id);
   }
 
+  @ApiOperation({ summary: 'Изменить информацию о спикере' })
+  @ApiResponse({ status: 200, type: SpeakersResponseType })
+  @ApiConsumes('multipart/form-data')
+  @ApiBody({ type: SpeakerUpdateBody })
+  @Roles('модератор')
+  @UseGuards(RolesGuard)
   @Put(':id')
-  update(
+  @UseInterceptors(FileInterceptor('file'))
+  async update(
     @Param('id', ParseIntPipe) id: number,
     @Body() speakerDto: CreateSpeakerDto,
+    @Req() request: Request,
+    @UploadedFile() file: Express.Multer.File,
   ) {
-    return this.speakersService.updateSpeaker(id, speakerDto);
+    const token = request.cookies.meetups_access_token.token;
+    const user = this.jwtService.verify(token, { secret: 'SECRET' });
+
+    await this.minioService.createBucketIfNotExists();
+
+    const speaker = await this.speakersService.updateSpeaker(
+      user.id,
+      id,
+      speakerDto,
+      file && file.originalname,
+    );
+
+    if (file) {
+      await this.minioService.deleteFile(
+        speaker.avatarImg.split('meetups-app')[1],
+      );
+      await this.minioService.uploadFile(`/speakers/${id}/`, file);
+    }
+
+    return speaker;
   }
 }
